chore: Make caching a little more intuitive
By default, each Dockerfile command is cached.

We should be leveraging the Gemfile.lock file in order to prevent
generating one differently at build time. There is a chance that a
developer generated a Gemfile.lock and pushed it up to GitHub, but when
we build/deploy the image, we wind up using different packages than what
is specified in the (tracked) Gemfile.lock.